% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation}
\alias{correlation}
\title{Create a correlation object}
\usage{
correlation(x, y, rho, ...)
}
\arguments{
\item{x}{The name of the first variable}

\item{y}{The name of the second variable}

\item{rho}{The correlation between the two variables}

\item{...}{Additional arguments to specify data frame names and factor names
See details for more information on the additional arguments.}
}
\value{
A list containing the correlation information
}
\description{
A helper function to create a correlation object
}
\details{
This function is a helper function to create a correlation
object that can be used to specify correlations between variables
when synthesising data using the \code{\link{synthesise_data}} function.
Additional Arguments:
\itemize{
\item df_name: The name of the data frame containing both variables
\item df_name.x: The name of the data frame containing the first variable
\item df_name.y: The name of the data frame containing the second variable
\item factor_name.x: The name of the factor variable for the first variable
\item factor_name.y: The name of the factor variable for the second variable
}
}
\examples{
 correlation("age", "bmi", 0.5)
}
