% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_variables}
\alias{filter_variables}
\title{Filter Variables}
\usage{
filter_variables(dfs, variables)
}
\arguments{
\item{dfs}{A list of data frames}

\item{variables}{A vector of variable names}
}
\value{
A list of data frames with only the specified variables
}
\description{
Filters a list of data frames to only include specified variables
}
\details{
This function filters each data frame in the input list to only include the specified variables.
}
\keyword{internal}
