% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal_distributions.R
\name{get_marginal_distributions}
\alias{get_marginal_distributions}
\title{FUNCTION_TITLE}
\usage{
get_marginal_distributions(df, variables = c(), ignore_na = TRUE, print = TRUE)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{variables}{PARAM_DESCRIPTION, Default: c()}

\item{ignore_na}{PARAM_DESCRIPTION, Default: TRUE}

\item{print}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
   marginal_distributions <- get_marginal_distributions(
     IST,
     variables <- c(
       "SEX",
       "AGE",
       "ID14",
       "RSBP",
       "RATRIAL"
     )
   )
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_all}}
}
