% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_missing_variables}
\alias{get_missing_variables}
\title{Get Missing Variables}
\usage{
get_missing_variables(dfs, variables)
}
\arguments{
\item{dfs}{A list of data frames}

\item{variables}{A vector of variable names}
}
\value{
A vector of missing variable names
}
\description{
Returns a list of missing variables from a list of data frames
}
\details{
This function checks if each variable in the input vector is present in any of the data frames.
}
\keyword{internal}
