% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise_data.R
\name{import_cor_matrix}
\alias{import_cor_matrix}
\title{Import a correlation matrix (removed)}
\usage{
import_cor_matrix(...)
}
\arguments{
\item{...}{Ignored, retained for backwards compatibility.}
}
\value{
No return value, always throws an error.
}
\description{
This function has been removed. Correlations should now be
specified using the \code{\link{correlation}} function.
}
\details{
Previously this function imported a correlation matrix from a
csv file. Correlations are now supplied directly to
\code{\link{synthesise_data}} as a list of objects created with the
\code{\link{correlation}} function.
}
\examples{
 try(import_cor_matrix())
}
\seealso{
\code{\link{correlation}}
}
