% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IST}
\alias{IST}
\title{IST Dataset}
\format{
A data frame with 19435 rows and 112 columns:
\describe{
\item{AGE}{#}
\item{CMPLASP}{#}
\item{CMPLHEP}{#}
\item{CNTRYNUM}{#}
\item{COUNTRY}{#}
\item{DALIVE}{#}
\item{DALIVED}{#}
\item{DAP}{#}
\item{DASP14}{#}
\item{DASPLT}{#}
\item{DAYLOCAL}{#}
\item{DCAA}{#}
\item{DCAREND}{#}
\item{DDEAD}{#}
\item{DDEADC}{#}
\item{DDEADD}{#}
\item{DDEADX}{#}
\item{DDIAGHA}{#}
\item{DDIAGISC}{#}
\item{DDIAGUN}{#}
\item{DEAD1}{#}
\item{DEAD2}{#}
\item{DEAD3}{#}
\item{DEAD4}{#}
\item{DEAD5}{#}
\item{DEAD6}{#}
\item{DEAD7}{#}
\item{DEAD8}{#}
\item{DGORM}{#}
\item{DHAEMD}{#}
\item{DHH14}{#}
\item{DIED}{#}
\item{DIVH}{#}
\item{DLH14}{#}
\item{DMAJNCH}{#}
\item{DMAJNCHD}{#}
\item{DMAJNCHX}{#}
\item{DMH14}{#}
\item{DNOSTRK}{#}
\item{DNOSTRKX}{#}
\item{DOAC}{#}
\item{DPE}{#}
\item{DPED}{#}
\item{DPLACE}{#}
\item{DRSH}{#}
\item{DRSHD}{#}
\item{DRSISC}{#}
\item{DRSISCD}{#}
\item{DRSUNK}{#}
\item{DRSUNKD}{#}
\item{DSCH}{#}
\item{DSIDE}{#}
\item{DSIDED}{#}
\item{DSIDEX}{#}
\item{DSTER}{#}
\item{DTHROMB}{#}
\item{DVT14}{#}
\item{EXPD14}{#}
\item{EXPD6}{#}
\item{EXPDD}{#}
\item{FAP}{#}
\item{FDEAD}{#}
\item{FDEADC}{#}
\item{FDEADD}{#}
\item{FDEADX}{#}
\item{FDENNIS}{#}
\item{FLASTD}{#}
\item{FOAC}{#}
\item{FPLACE}{#}
\item{FRECOVER}{#}
\item{FU1_COMP}{#}
\item{FU1_RECD}{#}
\item{FU2_DONE}{#}
\item{H14}{#}
\item{HOSPNUM}{#}
\item{HOURLOCAL}{#}
\item{HTI14}{#}
\item{ID14}{#}
\item{ISC14}{#}
\item{MINLOCAL}{#}
\item{NCB14}{#}
\item{NCCODE}{#}
\item{NK14}{#}
\item{OCCODE}{#}
\item{ONDRUG}{#}
\item{PE14}{#}
\item{RASP3}{#}
\item{RATRIAL}{#}
\item{RCONSC}{#}
\item{RCT}{#}
\item{RDATE}{#}
\item{RDEF1}{#}
\item{RDEF2}{#}
\item{RDEF3}{#}
\item{RDEF4}{#}
\item{RDEF5}{#}
\item{RDEF6}{#}
\item{RDEF7}{#}
\item{RDEF8}{#}
\item{RDELAY}{#}
\item{RHEP24}{#}
\item{RSBP}{#}
\item{RSLEEP}{#}
\item{RVISINF}{#}
\item{RXASP}{#}
\item{RXHEP}{#}
\item{SET14D}{#}
\item{SEX}{#}
\item{STRK14}{#}
\item{STYPE}{#}
\item{TD}{#}
\item{TRAN14}{#}
...
}
}
\usage{
IST
}
\description{
The International Stroke Trial Dataset
}
\details{
Obtained from
Sandercock, Peter; Niewada, Maciej; Czlonkowska, Anna. (2011).
International Stroke Trial database (version 2), [dataset].
University of Edinburgh. Department of Clinical Neurosciences.
https://doi.org/10.7488/ds/104.
Under ODC-by licence
}
\references{
\doi{10.7488/ds/104}
}
\author{
Sandercock P et al. \email{Peter.Sandercock@ed.ac.uk}
}
\keyword{datasets}
