% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RESIDE.R
\name{print.summary.RESIDE}
\alias{print.summary.RESIDE}
\title{print.summary.RESIDE}
\usage{
\method{print}{summary.RESIDE}(x, ...)
}
\arguments{
\item{x}{an object of class summary.RESIDE}

\item{...}{Other parameters currently none are used}
}
\value{
The summary.RESIDE object, invisibly.
Called to print to the terminal.
}
\description{
S3 override for print summary.RESIDE
}
\details{
S3 Override for summary.RESIDE Class, prints the overall
summary followed by a table summarising each data frame.
}
\seealso{
\code{\link{summary.RESIDE}}
}
