% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RESIDE.R
\name{summary.RESIDE}
\alias{summary.RESIDE}
\title{summary.RESIDE}
\usage{
\method{summary}{RESIDE}(object, ...)
}
\arguments{
\item{object}{an object of class RESIDE}

\item{...}{Other parameters currently none are used}
}
\value{
An object of class summary.RESIDE, a list containing
\code{overall}, a data frame of the overall summary, and
\code{data_frames}, a data frame with a row for each data frame.
}
\description{
S3 override for summary RESIDE
}
\details{
S3 Override for RESIDE Class, a higher level summary than
\code{\link{print.RESIDE}}. For each data frame it gives the number of
rows, subjects and variables, the number of each type of variable, the
number of date variables and the number of variables with missing data.
}
\examples{
summary(
  get_marginal_distributions(
    IST,
    variables = c(
      "SEX",
      "AGE",
      "ID14",
      "RSBP",
      "RATRIAL"
    )
  )
)
}
\seealso{
\code{\link{print.RESIDE}}
}
