% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise_data.R
\name{correlation}
\alias{correlation}
\title{Create a correlation object}
\usage{
correlation(x, y, rho)
}
\arguments{
\item{x}{The name of the first variable}

\item{y}{The name of the second variable}

\item{rho}{The correlation between the two variables}
}
\value{
A list containing the correlation information
}
\description{
A helper function to create a correlation object
}
\details{
This function is a helper function to create a correlation
object that can be used to specify correlations between variables
when synthesising data using the \code{\link{synthesise_data}} function.
}
\examples{
 correlation("age", "bmi", 0.5)
}
