% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise_data.R
\name{export_empty_cor_matrix}
\alias{export_empty_cor_matrix}
\title{Export an empty correlation matrix}
\usage{
export_empty_cor_matrix(
  marginals,
  folder_path = ".",
  file_name = "correlation_matrix.csv",
  create_folder = TRUE
)
}
\arguments{
\item{marginals}{The marginal distributions}

\item{folder_path}{Folder to export to, Default: '.'}

\item{file_name}{(optional) file name, Default: 'correlation_matrix.csv'}

\item{create_folder}{Whether the folder should be created, Default: TRUE}
}
\description{
A function to export a correlation matrix with
the required variables as a csv file
}
\details{
This function will export an empty correlation matrix
as a csv file, it will contain all the necessary variables including
dummy variables for factors.
}
\examples{
\dontrun{
if(interactive()){
 marginals <- get_marginal_distributions(IST)
 export_empty_cor_matrix(marginals)
 }
}
}
\seealso{
\code{\link[simstudy]{genCorMat}}
\code{\link[rio]{export}}
}
