% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESIDE-package.R
\docType{package}
\name{RESIDE-package}
\alias{RESIDE}
\alias{RESIDE-package}
\title{RESIDE: Rapid Easy Synthesis to Inform Data Extraction}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Assists researchers with planning analysis prior to obtaining data from Trusted Research Environments (TREs), also known as safe havens. Marginal distributions of one or more related data frames can be exported from a TRE and imported elsewhere, where data can be synthesised from them, with or without user specified correlations, by sampling from a multivariate cumulative distribution (copula). The International Stroke Trial (IST) is included as an example dataset under the ODC-By licence, Sandercock et al. (2011) \doi{10.7488/ds/104}, Sandercock et al. (2011) \doi{10.1186/1745-6215-12-101}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The RESIDE Package

This work was supported by the UKRI Strength in Places Fund (SIPF)
Competition, #' project number 107140.
The project title is SIPF The Living Laboratory driving economic growth
in Glasgow through real world implementation of precision medicine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hehta.github.io/RESIDE/}
}

}
\author{
\strong{Maintainer}: Ryan Field \email{ryan.field@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-4424-9890}{ORCID})

Authors:
\itemize{
  \item Ryan Field \email{ryan.field@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-4424-9890}{ORCID})
  \item David McAllister \email{david.mcallister@glasgow.ac.uk} (\href{https://orcid.org/0000-0003-3550-1764}{ORCID})
}

Other contributors:
\itemize{
  \item Claudia Geue \email{claudia.geue@glasgow.ac.uk} (\href{https://orcid.org/0000-0003-2243-0733}{ORCID}) [contributor]
}

}
