% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise_data.R
\name{synthesise_data}
\alias{synthesise_data}
\alias{synthesize_data}
\title{Synthesise data from marginal distributions}
\usage{
synthesise_data(marginals, covariance_matrix = NULL, ...)

synthesize_data(marginals, covariance_matrix = NULL, ...)
}
\arguments{
\item{marginals}{an object of class RESIDE}

\item{covariance_matrix}{Covariance Matrix not yet implemented, Default: NULL}

\item{...}{Additional parameters currently none are used.}
}
\value{
a data frame of simulated data
}
\description{
Allows the synthesis of data from marginal
distributions obtained from a Trusted Research Environment (TRE)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{is}}
}
